+++ src/details/tests.rs
@@ -17,23 +17,23 @@ extern crate rand;
 use std::collections::HashMap;
 use std::str::FromStr;
 
-use serde_json;
-use rand::{Rng, SeedableRng, StdRng};
+use bitcoin::blockdata::transaction::SigHashType;
 use bitcoin::network::constants::Network;
 use bitcoin::util::privkey::Privkey as RawPrivateKey;
-use bitcoin::blockdata::transaction::SigHashType;
+use rand::{Rng, SeedableRng, StdRng};
 use secp256k1::key::PublicKey as RawPublicKey;
 use secp256k1::Secp256k1;
+use serde_json;
 
-use exonum::helpers::{self, Height};
 use exonum::crypto::Hash;
-use exonum::storage::StorageValue;
 use exonum::encoding::serialize::{encode_hex, FromHex};
 use exonum::encoding::Field;
+use exonum::helpers::{self, Height};
+use exonum::storage::StorageValue;
 
+use details::btc;
 use details::btc::transactions::{sign_tx_input, verify_tx_input, AnchoringTx, BitcoinTx,
                                  FundingTx, TransactionBuilder, TxKind};
-use details::btc;
 use details::btc::HexValueEx;
 
 pub fn dummy_anchoring_tx(redeem_script: &btc::RedeemScript) -> AnchoringTx {
@@ -171,6 +171,22 @@ fn test_segwit_txid() {
 
     assert_eq!(tx.id().to_string(), txid_hex);
     assert_eq!(tx.wid().to_string(), wtxid_hex);
+}
+
+#[test]
+fn test_segwit_tx_builder() {
+    // Testnet transaction with id 6f3c41d81bfa04b6a96501344ddff630188ccf48c2fd4cf14cf02c3574f29844
+    let tx = FundingTx::from_hex("0200000000010103397c68945f02ced8dd61c4bfec516f585ff02b368ba6728bd112ef34d9b8541f00000017160014f3b1b3819c1290cd5d675c1319dc7d9d98d571bcfeffffff02a08601000000000017a914a144400568670200cf28be347be9126654cee14387d8f7d4000000000017a914b583eb93b9f86a878a90e0f16e1114c4b08ba8a6870247304402202cb276925bb2e932c2c5b0636c2dd6c360ae3178f1103e1d751b1caf12ab39ae0220387bd6e6a82a80da5a61b32a96653b2f2f6746fe58fd8a444db02ad4bce1bfb2012103150514f05f3e3f40c7b404b16f8a09c2c71bad3ba8da5dd1e411a7069cc080a0e8b81300").unwrap();
+    let anchoring_tx = TransactionBuilder::with_prev_tx(&tx, 0)
+        .fee(1000)
+        .payload(Height(0), Hash::zero())
+        .send_to(btc::Address::from_str("2NFGToas8B6sXqsmtGwL1H4kC5fGWSpTcYA").unwrap())
+        .into_transaction()
+        .unwrap();
+    assert_eq!(
+        anchoring_tx.0.input[0].prev_hash.to_string(),
+        "6f3c41d81bfa04b6a96501344ddff630188ccf48c2fd4cf14cf02c3574f29844",
+    );
 }
 
 #[test]
@@ -705,12 +721,12 @@ mod rpc {
     use bitcoin::util::base58::ToBase58;
     use bitcoinrpc;
 
-    use exonum::helpers::{self, Height};
     use exonum::crypto::{hash, Hash};
+    use exonum::helpers::{self, Height};
 
-    use details::rpc::{AnchoringRpcConfig, BitcoinRelay, RpcClient};
-    use details::btc::transactions::{AnchoringTx, FundingTx, TransactionBuilder};
     use details::btc;
+    use details::btc::transactions::{AnchoringTx, FundingTx, TransactionBuilder};
+    use details::rpc::{AnchoringRpcConfig, BitcoinRelay, RpcClient};
 
     fn anchoring_client() -> RpcClient {
         use std::env;