+++ src/details/btc/mod.rs
@@ -16,7 +16,6 @@
 
 mod types;
 mod redeem_script;
-mod address;
 mod private_key;
 mod public_key;
 pub mod payload;